requirements.txt
Django==1.11.5
django-environ==0.4.4
djangorestframework==3.6.4
elasticsearch-dsl==5.3.0
psycopg2==2.7.3.1
